Touch mouse in air with hover and click zones.
The interaction box is mapped to the screen area to control the mouse. The
half closest to the user is the hover zone (represented by a black ring)
and the other is the left click zone (with a red ring). The ring around the
mouse is smallest at the boundary between the two zones.